adaptation to types to deal with both 32 and 64 archs, need testing

+++ config/armv8a/chips/bcm2837/chipsupport.c
@@ -4,15 +4,15 @@
  * Created Date: Wednesday November 4th 2020
  * Author: Ronan (ronan.lashermes@inria.fr)
  * -----
- * Last Modified: Thursday, 6th May 2021 10:31:26 am
+ * Last Modified: Wednesday, 26th May 2021 2:43:30 pm
  * Modified By: Ronan (ronan.lashermes@inria.fr>)
  * -----
  * Copyright (c) 2020 INRIA
  */
 
-#include "chipsupport.h"
+#include "support.h"
 
-void security_domain_switch(unsigned int domain_id) {
+void security_domain_switch(WORD domain_id) {
     
 }
